api/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
alembic==1.13.2
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
backoff==2.2.1
black==24.8.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.0
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
executing==2.0.1
fastapi==0.112.0
filelock==3.15.4
FlashRank==0.2.9
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.32
langchain-openai==0.1.21
langchain-postgres==0.0.9
langchain-text-splitters==0.2.2
langgraph==0.2.4
langgraph-checkpoint==1.0.3
langsmith==0.1.99
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.4.2
numpy==1.26.4
onnxruntime==1.18.1
openai==1.40.6
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.3
pexpect==4.9.0
pgvector==0.2.5
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==5.27.3
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypdfium2==4.30.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rank-bm25==0.2.2
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.37.2
streamlit==1.37.1
sympy==1.13.2
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
yarl==1.9.4
